Change Watch Face Push API exceptions to no longer depend on wear-sdk.
This allows third-party developers to write tests of their apps that
use the API using mocks, as they can now simulate the API to throw
exceptions as required.

Additionally, API methods are now annotated with @Throws so that
frameworks like Mockito can correctly assert that the methods are
throwing the expected exceptions.

Change-Id: Ibbe4b07c8c53e1317e9c841d58bc357e5996b8b1
Bug: 3620529