Fix window null even after getting root in active window
AccessibilityNodeInfo#getWindow returns null even after
UiAutomation#rootInActiveWindow returned a non-null node. This is most
likely due to a race condition so we can only retry when that happens.
Note that this particular case is also very hard to test since it's
a race condition.

Bug: 417046391
Test: ./gradlew test:uiautomator:uiautomator:cC
Relnote: Fixed waitForStableInActiveWindow throwing NPE in some cases
Change-Id: Ibf50f93af389bc435cf3d2534a705d91d200b238