Move NavDisplay to common sourceset
The purpose of this change is to move the NavDisplay class out of the
android sourceSet and into the common sourceset.

This required the NavDisplay to stop using the `PredictiveBackHandler`
and move to an internal implementation that is backed by the
new AndroidX NavigationEvent library.

This also adds some internal hooks for compose that should eventually
be moved out into a NavigationEvent Compose artifact.

RelNote: N/A
Test: added new NavEventTest and tested in sample app
Change-Id: Ied9d5f544f8b1ab988ca4933642da0468edd9ec9