Remove OpenOnPhoneDialogDefaults.iconContentDescription which is no longer needed and add modifier to ConfirmationDialogDefaults SuccessIcon and FailureIcon
We should ignore semantics for icons and announce the text only in OpenOnPhoneDialog - so it is clearer to remove the contentDescription parameter from the OpenOnPhoneDialogDefaults.icon() and also remove OpenOnPhoneDialogDefaults.iconContentDescription.

ConfirmationDialogDefaults was missing the modifier parameter in the SuccessIcon and FailureIcon composables, so adding those here.

Bug: 409702471
Test: Tested manually
Relnote: "OpenOnPhoneDialog now announces only the message text under TalkBack, and skips the icon semantics. The OpenOnPhoneDialogDefaults have been updated to remove the iconContentDescription and the contentDescription parameter from icon(). Also, ConfirmationDialogDefaults now has modifier parameters on SuccessIcon and FailureIcon composables."
Change-Id: Id2ae2a01db6a94ce63a448693922d92beac10c01