Set up testapp to access internal datastore-core APIs
This commit sets up the testapp to allow accessing the internal APIs of
the datastore-core library for testing purposes. This is done by
creating a new configuration 'friends' in the build.gradle file and
adding the datastore-core project to it. The implementation
configuration then extends from 'friends'. This removes the need for
suppressing `INVISIBLE_REFERENCE` and `INVISIBLE_MEMBER` in the test
code.

TESTED: MultiProcessDataStoreIpcTest now passes!

BUG: 407740332

Change-Id: If4b90b3311ae4e67a98c793210175df3246f6767