Implementation of layout update state platform data.
This CL changes the layout update from boolean to int. This will allow developers to handle error cases in the layout update process.

Bug: 337187364
Test: Manual and unit tests
Relnote: We've introduced a new API that provides updates on the layout's status. This API uses a DynamicType with three distinct states to indicate when a new layout is being fetched.
Change-Id: I31ed99cf758ed450598d66606db7314a95ad151f