Fix dispatching onForgotton empty groups and elided groups
When groups are removed from the slot table the remember manager
is tasked with ensurin that the `RememberObserver` events are
dispatched in the correct order. With group eliding it could not
rely strictly on the order of the slots in the slot table as slots
added after a child group was added are appended to the end of the
groups slots. The previous code tried to provide enough information
for the remember manager to reorder the `onForgotton` calls. This
information was incomplete and the order was sometime wrong as
demonstrated by b/417450723. This change the slot table itself
calculates the remember order an always adds objects in the
remember order to the remember manager which is the order they
would occur if group eliding is disabled.

Relnote: """Fixed an ordering issue with the dispatching of
`onForgotton` that could, for example, cause the `onDispose`
of disposable effects to execute in the wrong order."""
Fixes: b/417450712
Test: ./gradlew :composer:r:r:test

Change-Id: Ic1c9127c3faf980e373eaa2c162a6cc49d55ff28