Focus Interop Rewrite
The isViewFocusFixEnabled flag fixes multiple bugs, by
always returning true when requestFocus() is called, and
re-directs focus to another view if compose does not have
focusable content. This CL changes this behavior and instead
prevents AndroidComposeView from being considered for focus
during focus search by overriding addFocusables() and
getFocusedRect().

This CL re-implements the code covered by isViewFocusFixEnabled
and breaks it up into three separate flags:

1. isBypassUnfocusableComposeViewEnabled
This flag enables the code path that causes AndroidComposeView
to be bypassed from FocusSearch so that it is not considered
for focus when it does not have focusable content. This is the
main part of the CL.

2. isPre26FocusFinderFixEnabled
Prior to API26, FocusFinder.focusSearch did not have loop
detection, and so one dimensional focus searches could result
in a loop. This flag enables a custom implementation of focus
search to cover cases where we have views behind the compose
view.

3. isIgnoreInvalidPrevFocusRectEnabled
When we receive a call to requestFocus with an invalid
value for the previouslyFocusedRect parameter, the view
system ignores the previouslyFocusedRect and assigns
focus to some sub-view. We added this same behavior to
Compose(b/388590015).

The CL also replaces moveFocus calls in interop tests with key
events as this is more realistic since the view system doesn't
have a moveFocus API.

Note: Before reverting this CL consider turning the
isBypassUnfocusableComposeViewEnabled and
isIgnoreInvalidPrevFocusRectEnabled flags off.

Bug: 388590015
Bug: 406327273
Bug: 369256395
Bug: 378570682
Bug: 389994198
Bug: 391378895
Test: Updated existing tests.
Relnote: "Added a flag that allows us to ignore the previouslyFocusedRect
passed to requestFocus() for parity with the view system"
Relnote: "Added a flag that allows us to exclude the compose view in
response to addFocusables() when compose does not have any focusable
content"
Relnote: "Added a flag that enables a custom implementation of focus search
for API<26 which does not have loop detection"

Change-Id: I3a686d20fc7cc4467e00d94338f962589830b542