Fix Size.isEmpty() corner cases
The previous version was 12 instructions and very clever, but didn't cover all cases (as demonstrated by the new unit tests).

The new version is 16 instructrions (still branchless), but is a lot easier to understand. I couldn't find a way to get it to be < 16 and I don't think this code path matters enough to justify it.

Fixes: 417259183
Change-Id: Id7098d7a273ed51d74c6f781f39112b714ac556f