Map unsupported Device Type to Unkwown.
We are adding new Device Types to the Framework. In the scenario where an app is using the new Jetpack SDK but the device has an older version of the Framework, saving records with new Device Type would cause an IllegalArgument exception.
To avoid this scenario, we can map Device Types not supported by the Platform as Unknown.

Test: ./gradlew :health:health-connnect-client:test
Bug: 416343576
Change-Id: I9c5073d673d4260b9e1f4504872877875f960fba