WebViewBuilder: Add javascript interface support and switch to allowlist
The big headliner in this change is that we will no longer have a
"policy" but rather have an allowlist that developers can attach.
The rationale here is that only having allowlists
greatly simplifies our API and expected behavior.

The one nuance is that for the JS interface, this could end up with
two sets with the same origin pattern and js interface name.
To simplify this, we will just not allow the same interface
name to be re-used. That is validated for in Chromium.

Test: WebViewBuilderTest
Bug: 383099115
Change-Id: I178121a38e66bd10dae518358a89492fdaa64c18