Revert^2 "Allow Dialogs to draw behind Display Cutouts"
Bug: 416090710

This reverts commit ada20fcd54a636033362331f8339d42615118d3b.

Reason for revert: ktformat change. Will re-run ktformat for this commit

The code was changed to require a new ktFormat between the time
this was presubmit passed and the time it was committed. This
caused the build to fail when it was committed. The CL was then
immediately reverted. This CL reinstates the original CL with
ktFormat run.

Test: same tests as first checkin

Change-Id: I6f14405059defdd98afaf0077d7baf7576c6e9e9