Change the order when applying transformations to M3 components.
Before, the transformations applied to Wear Material3 components where
done in the order: background painter, container transformation, content
transformation.
Now, the first 2 are inverted, and we apply any passed in modifier
between them, so it is affected by container transformations.
This fixes cases like using a placeholder shimmer effect with elements
in a TLC using the transformation paramter.
Relnote: "Change the order in which we apply the effect of
SurfaceTransformation parameters"
Test: Manual
Bug: 396679168

Change-Id: I786cfe1468d88d9c0f36b200d69362b46abdad55