Gesture pickup additional fix
When adding the gesture pick up feature a regression was introduced. If the gesture changes axis mid-way, the system will think it was not consumed and will try to pick it up, canceling the ongoing gesture. In this fix, once a gesture starts on a given direction we will continue consuming drag events even if they completely on the cross axis orientation, which is in line with the previous behavior. The changes are protected by the existing flag DragGesturePickUpEnabled.

Test: Added new test
Fixes: 416832576
Change-Id: I8c5f2b9ddc1824d4a5bbce9666128f0c1ee4dacc