Add hidden scroll fragments and scrollable Ad Types
Add hidden fragments with view-based and Compose-based UI of a
scrollable container rendering the remote content as the first item
while the rest of the items are local text fields.

The content of first item which is remotely rendered is also a scrollable
container of text fields, it uses newly added AdTypes which are scrollable
containers with an option to either allow or disallow the client app
from scrolling.

These fragments are created to be used to test user interaction in
upcoming CLs.

Bug: 420597985
Test: None, the added fragments are not in use.

Change-Id: I8a2c4fbe1efbe14582a11b0a6d757cccf7697279