Use the unified way of determining data class copy
After https://youtrack.jetbrains.com/issue/KT-72722,
`data` class `copy` has the primary constructor as source PSI
(so as to track it as an explicit constructor usage).
Previously, it pointed to the enclosing class (`KtClass`), and
checking `isData` on the UMethod itself accidentally worked.
The correct way is to check `data` modifier on the containing class,
and the detector already has and uses it on `visitMethod`.
This commit changes the logic in `visitVariable` to use the same logic.

Bug: 353980920
Test: unit test re-enabled
Change-Id: I088a4a0279a01194f6321bf4f28387fc234027db