Rearrange CameraControl integration tests
Move focus metering tests from CameraControlDeviceTest to
FocusMeteringDeviceTest and make them only run on lab devices so that
the focus and metering can converged correctly.

Also make CameraControlDeviceTest to run on any one of the available
camera since the lens facing of the camera is not important for the
tests.

Bug: b/384850020
Test: ./gradlew camera:integration-tests:build
Change-Id: I0992f825756acf31bcd0dcd7bfc7fc3705e1571f