Recomposer: remove incorrect resetting of `snapshotInvalidations`
This was causing occasional cases of missed recompositions when a
snapshot change would happen at an unfortunate moment and end up
ignored. Specifically, it was affecting Molecule users using
`RecompositionMode.Immediate`, which doesn't coalesce pending
recompositions until the next screen frame (cashapp/molecule#416).

The included tests use two separate compositions, although this isn't
a prerequisite for the issue to manifest itself. The linked downstream
bug uses a single Molecule flow (so a single Recomposer with a single
Composition), but that's harder to reproduce reliably.

This regression was originally introduced in
59526cdd "Fix potential dead-lock when disposing compositions",
when resetting of `snapshotInvalidations` was moved closer to reading
it under the more granular synchronized block, and the original line
(resetting it after iterating through compositions) was not removed.

Fixes: b/419527812
Test: RecomposerTestsJvm.recompositionOnConcurrentSnapshotInvalidation
Test: RecomposerTestsJvm.recompositionOnConcurrentSnapshotInvalidationStressTest
Relnote: """Fixed a bug affecting Molecule users where using
`RecompositionMode.Immediate` could cause missed recompositions."""

Change-Id: I9f3a9b3d879aef23a73eb4110f2efed6cca81dfa