Fix fling cancellation issue
Fix fling cancellation issue. We're restoring the fling continuation behavior only canceling the fling now if the node is detached.

Relnote: Introduced isFlingCancellationWithNestedScrollFixEnabled to fix an issue with fling propagation in nested scrolling. In this CL we are restoring the fling continuation behavior removed in aosp/3260391. We will still cancel the fling animation in case the child is removed from composition.

Test: Added new test

Fixes: 405910180
Fixes: 419049142
Fixes: 416784125

Change-Id: I467f4a8974a752d63bd3d6ab62e6eb751f5e2ccc