Remove DisposableEffect to clear backstack state
Since decorator states are now hoisted, navigation should not differentiate between the different ways that the backstack is modified, regardless of whether it was:

1. a mutable list that was modified
2. an immutable list that was swapped

It should merely reflect the current backstack/decorator state. To do that we need to stop clearing the state for backstacks that were swapped out.

Test: ./gradlew navigation3:navigation3-runtime:cC
Bug: 415076044
Relnote: "Navigation3 will no longer clear decorator states for backStacks that have been swapped out and replaced with another backStack instance."
Change-Id: I28a42b98cc811be99d8380b1cfceee02b42608ab