Relaxing error handling in the OpenXR update blocks for Anchor, Plane and Hand.
Setting the TrackingState to PAUSED when the JNI code cannot resolve the latest state (instead of producing an uncaught IllegalStateException). This will give the OpenXR runtime an opportunity to recover before the next update loop. As a future improvement, we can consider changing the value to STOPPED if the runtime cannot produce the state after a while.

Also updated the renderers in HelloAr to reflect TrackinState.PAUSED with 50% transparency for planes and anchors.

Bug: 414392547
Tests: On device with HelloAr
Change-Id: I8f6ca6dd495dde1c2a0203c131cee9a4358141e7