Fix for remeasure being skipped on a non placed child when it is measured in placement block
Previously the logic was skipping remeasure requests from the the non placed children if they were measured in the placement block. It is true, that such children can't affect the parent size, but they still can affect parent, for example if a sibling is using the measured size of the unplaced child as its position. The correct logic should only skip remeasure requests if the child was not measured at all by the parent.

Test: new tests in PlacedChildTest
Fixes: 239087806
Change-Id: Iddfefb9637c6d7019968212d23c31668bef06e6a