Fix missing InMemoryTracing events when zipTraceWithInMemoryEvents.enable is used
Zip InMemoryTracing instead of writing directly to the trace file. This
writes the InMemoryTracing and Perfetto-captured trace as two distinct
files, and then it zips them together to produce the final combined
trace.

This is only enabled when zipTraceWithInMemoryEvents.enable is set to
true to avoid issues with Studio not supporting the trace format.

Test: TrivialKotlinBenchmark // with and without zipping
Test: PerfettoTraceRuleTest // with and without zipping
Bug: 421955180
Change-Id: Id7cea9ac24368709d939490e872f20f4cbd86bb4