Fix calculations of adjusting maxLines of text in its available space
Previosly, we were using lineHeight set to the TextView only.
However, this has an issue, because lineHeight is respected only
on all lines except the last one. Last one (or only one in case
of single-line) is handled by Android, to make sure it has enough
space for tall glyphs.

Fix is to use line top and bottom positions to get each inidividual
line height and based on that see how many can fit in the available
space for height of text.

Fixes: 414353620
Test: Existing ones
Relnote: "Fix calculations for adjusting max lines in the renderer side."
Change-Id: I933bc3969a34c5de4eb320225f3f695db88db4ff