Catch IllegalStateExceptions and not report the error
When executing camera framework calls, the camera could be closed or
disconnected. In such cases, an IllegalStateException is thrown by the
framework. Since these exceptions can still occur at any time (e.g., we
may get notified of a disconnection after a delay), there is a practical
reason for us to catch these IllegalStateExceptions and not surface them
as a true graph error.

Bug: 420735918
Test: Tested with camera apps
Change-Id: I0a8ef3f76caa55f633107f555f6571d888682b98