Inline snapshot observation functions
Marked as inline functions used only in one places in the measure/layout system, mostly related to snapshot observation. Before this change there are 5 function calls needed only to run the observation, after the change there is only 1 function. According to benchmarks it doesn't have any noticeable effect on the benchmark numbers (like LazyListScrollingBenchmark), but the positive effect on the smaller stack trace is noticeable. First, it is easier to read the traces, second there is still some runtime overhead for each function call in a stacktrace.
Also, the observation lambdas are now preallocated for lookahead in the same way as in the main pass.

Test: looked at traces in benchmarks
Relnote: n/a
Change-Id: I6d8ea6ac4b3c54fd8ed9c03b77176db80cc8b3d8