Handle SecurityException when opening cameras
As of right now, our camera device manager checks whether the user has
camera permissions before attempting to open the camera. Since there is
a dedicated error for it already, in this CL we try to handle this by
removing the check, and instead handle it by catching the
SecurityException (which we already do). This CL also fixes an issue
where the noisy nature of framework signals could lead to unnecessary
retries when the last error we received is a SecurityException.

Bug: 419602855
Test: Tested with CameraPipe sample app with camera permission removed.
Change-Id: I925b831d0b2a9ed7a0d39a4ee6d263371f4c2b51