Simplify KMP AGP source jar publication.
We were creating a task that was never used. If you check the androidx target in postsubmit, you'll see that we never run sourceJarAndroidMain. Source jar publication was happening via the KMP-provided androidSourcesJar task instead
For KMP projects with an Android target, the `sourcesJar` configuration is not used by AGP for publishing. This change removes it from the publishing variants.

BUG: 420602018
TESTED: Ran publish task and the android sources jar was created with the correct content

Change-Id: Iaa050b88e2273c3a1527ec46fa02b7043d32c221