Fix complex interaction between pager and scaffold.
When paging between Screen scaffolds containing transforming lazy
columns and edge buttons, a DynamicHeightNode instance can be reused
without its lastMeasureHeight not reset to null, resulting in it not
informing its height through the callback, so no space was reserved for
the edge button. That was wrong and its fixed now.
Also, changed the ScreenScaffold to use SubComposeLayout so that
happens before we need to create the content, so we know the button
height and how much space to reserve on the first pass.
Relnote: Fixed complex interaction between pager and scaffold.
Bug: 419819320
Bug: 409501073
Test: Added

Change-Id: I946e3cd808e5416a235277d0626e18ac9edb7fca