Fix Matrix4 scale to be always positive or negative based on the determinant
The returned rotation is now always unscaled.

Since matrix decomposition is not exact without preserving the scale
separately, we return one of the valid decompositions using the rotation
matrix determinant to determine the sign for all scale components.

Negating a scale for two axes is equivalent to a 180 degree rotation
around the third one, so the rotation matrix is adjusted accordingly to
compensate for it.

Test: ./gradlew :xr:scenecore:scenecore:test :xr:runtime:runtime:test

Change-Id: I2305d13af2958211baecf0e89593ff97939d4d15