Add kotlin 2.2 target for additional projects in KGP patch
As a followup to aosp/3644764, there are some additional projects which need to target kotlin 2.2. to pass in the integration workflow (for the swiftExportClasspathResolvable configuration). Also, since kotlin 2.2.20 stable doesn't exist yet, the `kotlin22 = "2.2.20"` line added to the patch in the previous change wasn't working, so this adds a step to the workflow to replace it with the latest version being tested -- this should be removed once the stable verion exists.

Test: Made the same changes in a playground branch (https://github.com/androidx/androidx/compare/androidx-main...juliamcclellan:androidx:kotlin22-patch) and verified that the workflow passed (https://github.com/juliamcclellan/androidx/actions/runs/15441382369)
Change-Id: I59b6d4d0bdf38aa3a8ba8a1cb44eb9e7005dab87