Skip traversal early if there are no onGloballyPositioned modifiers in the subtree
Initially when we introduced onGloballyPositioned, I thought it is safer to always traverse instead of trying to keep track of the fact if we have some modifiers set in the subtree or not. But recently we introduced similar tracking for onPositioned callbacks (child.layoutDelegate.childrenAccessingCoordinatesDuringPlacement), so I decided to use similar techinique to optimize onGloballyPositioned traversal. As we are planning to push on users to migrate away from using onGloballyPositioned with new visibility apis, the goal is for this mechanism to add zero cost whenever there are no usages of onGloballyPositioned in the subtree. More usages of onGloballyPositioned will be migrated away, more perf gain this mechanism will provide us.

RectsInColumnBenchmark.first_layout[100]
Before:
timeNs            min 1,604,519.3,   median 1,858,452.1,   max 3,617,327.3
After:
timeNs            min 1,568,806.8,   median 1,830,009.4,   max 3,688,192.0

Test: existing tests
Change-Id: I8665ce51c32542fd099c2760c112ad0542754ce7