Exclude ComposableInferredTarget annotation from signature files
This annotation isn't important for compatibility, and may change when there are no other API changes.

Because ComposableInferredTarget is itself annotated @InternalComposeApi, it was considered a suppress compatibility annotation. Removing it removes @SuppressCompatibility unless a method is already annotated with another experimental annotation. This means these APIs previously weren't considered for compatibility checks, but now are.

Bug: 422450071
Test: ./gradlew checkApi
Relnote: N/A
Change-Id: I0f521071b890ca61ca6f56c03b94e53bb5e76270