Expose CameraX Feature Group APIs
 This CL also improves the documentation while making the APIs public as experimental.

RelNote:  "Added new APIs to set or query groups of high-quality features more intuitively and accurately as these features may not be supported as a combination across all devices. SessionConfig now contains new APIs to denote a group of features as either required or preferred for a camera session. The CameraInfo.isFeatureGroupSupported API allows to know if a feature group is supported before binding a SessionConfig. SessionConfig also allows setting a listener to know which features are finally selected when a session config is bound."

Bug: 422644009
Test: ./gradlew -p camera updateApi, ./gradlew docs
Change-Id: Ie4d60e3ef392ae7e5016cc9c102e5c72bb6a10fd