Fix for RectManager not adding an offset added by layout cooperation
Layout cooperation is a mechanism of automatically centering a child in a parent when the child is using the size not within the constraints
provided by the parent.
In the previous implementation we were reporting to rectManager the plain offset provided by the parent, without taking into account an additional offset Placeable will add to accommodate the layout cooperation. To fix it, I've moved the position reporting right into the outer node coordinator, to use the final offset.

Test: new test in OnGlobalRectChangedTest
Change-Id: Id8077b91937e4f0427c2579b6c394d727b3e2871