Elevation and FAB updates at the FloatingToolbar
To allow arbitrary shapes that are applied to the provided FAB via a
slot, we no longer lerp the shadow of the FAB content as the toolbar
expands or collapse. The lerping in now only done for the toolbar's
container only.

Test: Updated
Fixes: 423336922
Relnote: "Fix an issue where an arbitrary shape on a FAB that is
passed to a FloatingToolbar was not applying its shadow correctly."

Change-Id: Icdcc9be21d9c7951520f9a1cd1f6fe1d136f5320