Fix how increments for text autosizing in M3 are handled for scalability
Previously, we were calculating DP to SP separately on a base size
and then on the increments. Since increments can be negative, this
leads to issues when text is non scalable, as converter uses them
as positive values. Also, those calculations are in float numbers,
and they would have decimals, however, toInt() just removes all
decimals.

Fix is to get what size should be and then apply scalability and
use roundToInt() to better translate float to int, as TextView's
autosizing only accepts ints. This means, that even if the same
size can fit for the case when there is autosizing, vs when there
isn't, results might defer as that size could be bumped to a
higher rounded number and no longer fit.

Additionally, fix an issue with maxLines calculations in
ProtoLayoutInflater.

Bug: N/A
Test: Manually
Change-Id: I24beec948485ac59643086f39e0b4adbbb39b316