Make the initialization of OnBackPressedDispatcher and NavigationEventDispatcher lazy
Our current setup is like this:

```
class ComponentActivity {
    val onBackPressedDispatcher by lazy { ... }
    val navigationEventDispatcher: NavigationEventDispatcher by onBackPressedDispatcher::navigationEventDispatcher
}

class OnBackPressedDispatcher {
    val navigationEventDispatcher: NavigationEventDispatcher = NavigationEventDispatcher()
}

```

This makes `ComponentActivity.onBackPressedDispatcher` not lazy because it initializes a `navigationEventDispatcher$receiver` in ComponentActivity's constructor. This could lead to increased ComponentActivity instance creation time. We make it truely lazy in this CL by replacing the property delegate of ComponentActivity/ComponentDialog.navigationEventDispatcher with a custom getter.

This CL also makes OnBackPressedDispatcher.navigationEventDispatcher lazy by using the lazy delegate. This will delay the instance creation of NavigationEventDispatcher to after OnCreate (when `setOnBackInvokedDispatcher` is called).

Bug: 422044262
Test: existing tests still pass
Change-Id: I710e6cc0ef99009aff88c09a594e8c31bf204464