Updates surface focus and press states
Press overlay now has a minimum duration
Focus highlight is reimplemented with a runtime shader, to more closely match the design. Previously we were rotating a gradient without recreating it - this results in a non-uniform rotation in most cases as the gradient bounds will move outside of the surface bounds during rotation for asymmetrical shapes. Now we map the gradient to the rotated state, to make sure the gradient is fully within the bounds of the component during the rotation.

Bug: b/413428451
Bug: b/413429405
Test: SurfaceTest
Test: SurfaceScreenshotTest
Change-Id: If8d10914857f697a11de46ffe1a9bda4262c40c1