Move hiltViewModel() to new package and library
This change creates two new artifacts:
* androidx.hilt:hilt-lifecycle-viewmodel
* androidx.hilt:hilt-lifecycle-viewmodel-compose

The move is to remove the navigation dependency from the hiltViewModel() utility function that can be used outside of Navigation in a Compose app and similarly the factory function HiltViewModelFactory for non-compose without Navigation. This in turns enable the API to be used in a Compose app with Nav3 without bringing Nav2.

The factory function that was in hilt-navigation got moved to hilt-lifecycle-viewmodel. While the composable view model retrieving functions from hilt-navigation-compose got moved to hilt-lifecycle-viewmodel-compose. Dependencies on both hilt-navigation-* artifacts where updated to use the newly introduced artifacts, delegating to them.

Bug: 328104152
Test: HiltViewModelComposeTest + HiltNavigationViewModelComposeTest
Relnote: "The hiltViewModel() APIs for Compose have been moved to a new artifact (androidx.hilt:hilt-lifecycle-viewmodel-compose) and package (androidx.hilt.lifecycle.viewmodel.compose) such that they can be used without transitively depending on androidx.navigation."
Change-Id: Ifb2226c57e78d70a282f58bafaccabe8e3b51e7f