[TLC] Fix a bug with deletion of an anchored item and an item before it.
An `anchorItemKey` was added to `TransformingLazyColumnState` and related measurement classes. This allows the `TransformingLazyColumn` to correctly restore its scroll position when items are removed before the currently anchored item.

If the currently anchored item is deleted then the next item becomes an anchor item, but with an offset coerced at least at the top of the screen. This fixes a problem with disoriented scroll during item deletion.

Bug: 407575439
Test: TransformingLazyColumnContentPaddingMeasurementStrategyTest.kt, TransformingLazyColumnTest.kt
Change-Id: I841a80455ce4993ad58e513ee91d32d3be4fc128