Removed hard coded constants from tests
These tests hard coded constants from the Xr Extensions as they
were missing in the fakes. The fakes have been updated to have the
constants.

This change set removes the duplicate constants from the tests and
has them reference the constants as they are exposed in the Xr Extensions.

Bug=402408284
Tested=Ensured tests passed

Change-Id: Ie5f309241bf477abc71d7358cf68b6a1d47d8cc5