Fix animations with immutable backStack
Navigation keys backstack operations to backStack instance so that it can differentiate swapping backStack vs backStack operations (pop, navigate). However, this prevents immutable backstack operations (i.e. nav by making a copy of the current backstack and adding new key to the new backStack) from seeing animations because the `key` composable helper prevents the recomposition of the current destination.

Current solution is to remove the key helper and determine swap by comparing the root key between backStacks - it is considered a swap if the root key is different, or it is considered a backstack operation (pop/nav) if the rootkey is the same.

Note that this is a temporary best effort solution to re-enable animations again.

Bug: 423789042
Test: ./gradlew navigation:navigation3-runtime:cC
Test: ./gradlew navigation:navigation3-ui:cC
Change-Id: I0ccff01e30783aaf717e566167e089b5ed1e662c