Add clear focus on tap
Relnote: """Pointer downs with a mouse or touchpad in a ComposeView will
    now clear focus automatically if the pointer down doesn't occur in the bounds
    of the focused node. This results in a "tap-to-clear focus" UX that is
    more expected than current behavior when using pointer input devices.
    This behavior can be opt-ed out of with a new
    AbstractComposeView.isClearFocusOnPointerDownEnabled API.
    """

This adds additional pointer input handling in AndroidComposeView to clear focus on pointer downs
from a mouse or touchpad. This will clear focus from components that would otherwise keep focus
even when in touch mode.

This will be enabled by default when the newly added ComposeUiFlags.isClearFocusOnPointerDownEnabled
is flipped to true in the follow-up CL.
The intention is that this flag is flipped to true, then deleted, changing default behavior
for all Compose views.

This default behavior can be reverted on a per-AbstractComposeView basis by using the new
AbstractComposeView.isClearFocusOnPointerDownEnabled property, which will allow opting out of
this new behavior.

Bug: 282963174
Bug: 405813364

Test: new tests in FocusableTest.kt

Change-Id: I6322b0a63e80042f6a558104d0684443e5dc161f