Fix arrow key processing in BTF2
This change removes the preKeyEvent processing of arrow keys by the new TextField. Instead TextField handles all keys first, then reports back whether the received key was properly consumed.

When arrow keys do not get consumed e.g., when the cursor is at the end and the user presses RIGHT arrow key, this event is now gets delegated to the system.

Added a new KeyCommand for `DPAD_CENTER` key to open the keyboard on devices like TV remotes or gamepads.

Fix: 295666701
Fix: 410753791
Relnote: "Fixed a bug where certain hardware keyboards or TV remotes fail to move cursor in new TextFields using the directional keys, and instead switch focus to another composable. Added `ComposeFoundationFlag.isTextFieldDpadNavigationFixEnabled` to be able to disable this fix for the time being if it causes an undesired behavior in an irrecoverable way."
Test: gradle :compose:foundation:foundation:cAT
Test: gradle :compose:foundation:foundation:testUnitTest
Change-Id: Ie1922b03e8ec24a0b0d35995e407a3bcd09400a0