Changed uiautomation-shell server to accept multiple connections
This version changes the native shell server in main.c to accept
sequentially mutliple connection, instead of shutting down after the
first. It also removes the pipe and exposes the socket buffer via a
wrapper meant only to catch unexpected SocketExceptions. This new
version is also much more performant and it's able to execute a command
in less than 2ms (measure on a pixel 6a) vs the previous that is around
100ms.

Bug: 407998523
Test: ./gradlew test:uiautomator:uiautomator-shell:cC
Relnote: Changed uiautomation-shell server to accept multiple connections
Change-Id: I738a268686098d25ef8f36585cf797f281f67da0