Mark ValidateIntegrationPatches task as non-cacheable
Before this change the assumption was that only the patch change could
break the task, but it could also be changes to the androidx files.
We had a build break ab/13660509 where this happened and because of
caching it was not caught by test runners that run with remote cache
enabled.

Test: None
Change-Id: Ic2d9dd97693c65b4b52435144eb7e6607792dea7