recyclerview-selection: fix "double click on a selected file"
For real (not test) code, the important change is to MouseInputHandler
where code moved from onSingleTapUp to onSingleTapConfirmed.

Prior to this commit, onSingleTapUp would return true (and so prevent
onDoubleTap being called) if something was already selected, which broke
the "double click on a selected file" flow.

After this commit, we no longer override onSingleTapUp (which was passed
an MotionEvent.ACTION_UP event), only onDown, onSingleTapConfirmed and
onDoubleTap, which all get the MotionEvent.ACTION_DOWN event. This means
we can simplify by removing the mHandledOnDown and mHandledTapUp click
state tracking as well as the "Since we don't get a button state back
from UP events" comment.

This commit also cleans up the handling of right-click (secondary mouse
button) and middle-click (tertiary mouse button). It also removes some
no-longer-necessary Log.d and Log.v calls.

For test (not real) code, the important change is to
testDoubleClick_Opens, uncommenting a previously-broken test. The other
changes are minor.

Bug: 389814214
Bug: 427590351
Test: gradlew :recyclerview:recyclerview-selection:connectedAndroidTest
Change-Id: Iea74c19f158d60f55ba827c2ba53c623cbd293e8