Remove Saver from SwipeToReveal
When SwipeToReveal is paired with a lazy list, using rememberSaveable and a Saver object means that, once a revealed item is no longer visible, its state will be saved and it will be recreated as revealed once re-displayed on the screen. That breaks our UX spec that only one item should ever be swipe-to-reveal'd at a time.

Bug: 419189965
Test: Removed 'recreation' tests associated with the rememberSaveable and Saver object
Relnote: "Addressed bug where multiple revealed items could be shown with SwipeToReveal in lazy lists"
Change-Id: Ic7388e2236382bcdf86ea086d44f68a04a73baab