Add optional display name to PreviewParameter
This change adds a new `getDisplayName` method to the
`PreviewParameterProvider` interface. This method allows developers to
provide a custom display name for each parameter, which will be used in
the preview UI. If `null` is returned, the system will generate a
default name.

Fixes: 241699422
Test: Manual
Relnote: "Added a new getDisplayName method interface to allow custom
display names for preview parameter instances."

Change-Id: I19bdfbff9392179ebb9d033bcfb11d124c0f366e