Only capture nodes to a single sub composition
Before: if a node with an empty size had multiple sub compositions,
the node was moved to all of the sub compositions.

This is wrong. Instead continue to move empty size nodes to the
single sub composition it holds. If a node has multiple sub
compositions the node is kept in the current composition.

Bug: 425651406
Test: TODO
Change-Id: I41bc7a27d07db0e6d6316e16c62ec7767368983f