Repeat removal of Saver from SwipeToReveal (originally aosp/3667041), with the merge issue resolved that caused the build break.
When SwipeToReveal is paired with a lazy list, using rememberSaveable and a Saver object means that, once a revealed item is no longer visible, its state will be saved and it will be recreated as revealed once re-displayed on the screen. That breaks our UX spec that only one item should ever be swipe-to-reveal'd at a time.

Bug: 419189965
Test: Removed 'recreation' tests associated with the rememberSaveable and Saver object
Relnote: "Addressed bug where multiple revealed items could be shown with SwipeToReveal in lazy lists"
Change-Id: I1d4f68415a7f96e5bccc8619f447867a6e97f60a