Add e2e tests for PoolingContainer View CUJs
Adds a parameterised test class to verify PoolingContainer CUJs for View
Framework. The class utilises existing AutomatedEntToEndTest and a new
ReusableListViewHiddenFragment that is introduced to initialise a
RecyclerView with AdHolders as row items.

Tests verify remote content views are not destroyed on window detach,
but get destroyed when released by the PoolingContainer.

Parameters for the tests: ad format (banner/native), mediation option,
z-order.

Bug: 414333422, Bug: 422160411
Test: ReusableListTests

Change-Id: Id7a09aa7d16c6d31d77999577a71152232109ec3