Change decorator state onPop behavior
Navigation 3 no longer attempts to differentiate between backStack operations (pop/navigate) and backStack swapping (multiple backStacks). Instead, state management is separted into two layers:

1. If an entry is popped (regardless of whether it was an actual pop, or a swapped backStack that no longer contains a key that was in the previous backStack), the NavEntryDecorator.onPop for this entry will be called on the latest decorators. This means that if user wants to preserve the decorator states for the backStack keys that were swapped out, users must make sure to swap out the decorator states as well.

2. If a decorator was removed from the list of decorators that was passed to the NavDisplay (regardless of whether an individual decorator was popped, or the entire list of decorators was swapped out), nav 3 will no longer forcibly clear the states of the removed decorators. As users now own the hoisted decorated states, users will also have full control over whether and when they want to clear the states of removed decorators.

Test: ./gradlew navigation3:navigation3-runtime:androidConnectedTest
Bug: 428033667
Relnote: "A entry's state is now strictly based on the current list of decorators passed to the NavDisplay. This means that if you swap out backStacks in the case of mulitple backStacks, the decorator states should be swapped out as well to ensure that the old backStack's states are preserved. Otherwise, the states will be cleared as if the entries were popped (instead of a swap)."
Change-Id: I7a759e37580404cf9c3a35b1c2a8f18b3d20646e