Add adFormat as a test app launch param
Adding adFormat to FragmentOptions allows to explicitly specify the
format when launching the test app manually or in the automated tests.
adFormat options include "banner-ad" and "native-ad", with the former
being the default option.

adFormat is also added to AutomatedEndToEndTest - this will simplify
adding e2e tests for native ads/shared UI APIs.

Bug: 340471014
Test: adb shell am start -n androidx.privacysandbox.ui.integration.testapp/.MainActivity -e "fragment" "resize" -e "ad-format" "banner-ad"
Change-Id: Iddf68ec4186980ef657509d9f9602ead778a7cad