Simplify composing additional items in lookahead
This change simplifies retaining additional items in lookahead
pass by only composing items needed for lookahead in lookahead.
As such, the approach pass will compose additional items if
needed, and subsequently a make-up lookahead will be triggered
for these late-(sub)composed layouts.

An additional optimization is added to subcomposeLayout, such
that if a slot is no longer needed in lookahead but should
still be composed in approach. The node as well as the composition
will transfer ownership from lookahead to approach. This
1) avoids the unnecessary work of disposing the slot composed
in lookahead and again composing it in approach pass,
2) allows states (including animation states) to persist through
this ownership transfer, and
3) once approach pass is done using the slots that it owns, it
can safely and *promptly* dispose them without needing to trigger
lookahead.

Test: new tests added
Change-Id: Ic53f235c1f59dd1dcfa782824a7f2983facf2f13