Close SandboxedSdkView session after next frame commit
Instead of directly closing the session when detached from window, a
FrameCommitCallback is registered to ensure provider-owned UI is in sync
with client-owned UI.

Closing the session directly on detach is not reliable as the next
frame may not have rendered by this time, resulting in provider UI
becoming unavailable before everything else. An example of this is
a transition to another fragment.

Bug: 429380932
Test: SandboxedSdkViewTest, PoolingContainerTests, IntegrationTests
Change-Id: I62f7b4269016aa33ba386c664528eb184115e1f9