Do not update texture while configuring Session
In the whitebox DepthMapActivity, the depth texture is updated
asynchronously with calls to Session.configure(). This can result in
depth swapchain memory becoming invalidated while updating the texture.
Fix this by preventing updates to the depth texture while configuring.

In addition, performing some general refactoring of the activity, as
well as setting invalidated depth maps to null in the headset
implementation of the DepthMap interface.

Test: on Moohan device
Bug: 428227724

Change-Id: Ia7945fb3b8b1466ec2903182ecc0242a7c44f51a