Speculative workraround for robolectric test JVM crashes
The speculation is that the crash is due to a bug in Java C2 JIT
compiler that gets triggered by code generated by robolectric for the
android.icu.util.Calendar class.

Using -XX:CompileCommand=exclude tells runtime not to compile this
class, which means it will run slower, but hopefully not crash now.

https://docs.oracle.com/en/java/javase/12/docs/specs/man/java.html#:~:text=%2DXX%3ACompileCommand%3Dcommand%2Cmethod

Test: ./gradlew :wear:protolayout:protolayout-expression:testReleaseUnitTest
Bug: 428257656
Change-Id: I92a74aad7c1d108032b4498d72872eec5b476225