Add APIs for DataBridgeSdkProvider
Add registerKeyUpdates and unregisterKeyUpdates for DataBridgeSdkProvider. All the registration will be done via DataBridgeProxy.
The registerKeyUpdates will register keys which the caller is interested in getting updates, and the unregisterKeyUpdates API will unregister the callback and all further updates to the keys which were associated with the callbacks will not be sent

Bug: 423805466
Test: DataBridgeSdkProviderTest
Relnote: Add registerKeyUpdates and unregisterKeyUpdates for DataBridgeSdkProvider
Change-Id: Id01ce796c81f04b0bbff03df6931f841dec133f4